+++ theater/__init__.py
@@ -1,5 +1,4 @@
 from .base import *
 from .conflicttheater import *
 from .controlpoint import *
-from .frontline import FrontLine
 from .missiontarget import MissionTarget
